Staging of a major secp256k1 refactoring.
At the moment, the C and openCL build on the GPU work, however the openCL
build does not work for the CPU: apparently the Nvidia and intel openCL
compilers work differently. The cause of the CPU build failure is mixing
of __constant, __local and __global namespaces.

Fixing this by a straightforward renaming of types however
will break the C build, which is unacceptable. I have therefore
decided to make a rather involved macro-based fix that will allow all
three builds to work off the same code base.
The fix is going to be very laborious, so in this commit freezes the code
in a 2 working 1 broken build state. The macro-based fix of this problem
will be developed on a separate branch, where temporarily all 3 builds
will be broken, and merged into the master branch when all 3 builds work.