Removed non-trivial work from VERYFY_CHECK macro.
Furthermore the macro is now disabled.
Putting useful work in an assertion-style macro turned out
to cause big trouble.
When I disabled the macro, this broke the code.
If I am susceptible to making such mistake, despite being
already quite experienced with the code base, so will be
others. So, to avoid similar trouble in the future,
I will eventually remove the macro altogether.