Bug with crypto computations thoroughly documented, commenting out debug code.
As of right now, it remains unclear whether the bug is a 1) hardware bug,
2)Intel openCL driver bug, or 3) software bug. If 1) is the case,
then no action needs to be taken. If 2) is the case, we need to take action
by restricting builds and possibly filing a bug report with Intel.
If 3) is the case, we need to know and fix.

I propose no action until the bug is reproduced on more machines. The
branch https://github.com/blockchaingate/Kanban/tree/fixCryptoBuild remains
to document this known issue.