Refactored transfer of unsigned ints to the GPU.
Unsigned ints are now first converted with bitshift operators
to an array of 4 unsigned chars, then written to the kernel.
On the kernel side, unsigned integers are first cast to
an array of 4 unsigned chars, then read with bit shift operators.

This roundabout method of passing integers around appears
to be system independent.