Implemented dumpprivkey rpc call. Implemented security filter.
The security filter works at the
moment in the node layer.
RPC calls marked as unsecure are executed only if
the most secure network given in the arguments
has more relaxed security than testnet, testnet
excluded. At the moment, the security filter will let
regtest and testnetnodns through, but not
mainnet and testnet.