Added command to free https on remote machine.
May need security audit, for the following scenario.
Someone hijacks the remote machine. Then inserts himself
as an intermediate listening to our ports, passing all info
back to our system. In this scenario,
the attackers executable will not be killed with our standard remote
restart code. In this way, after a restart,
our system will no longer function because the
port is taken over by the harmful executable.

Now, when we kill all listening to the port, we kill the attacker's
executable and never find out about the breach (but
we have a running system). This is clearly not
desirable.